src/main/java/com/cs/rfq/decorator/extractors/AverageTradedPriceExtractor.java
package com.cs.rfq.decorator.extractors;

import com.cs.rfq.decorator.Rfq;
import org.apache.spark.sql.Dataset;
import org.apache.spark.sql.Row;
import org.apache.spark.sql.SparkSession;
import org.joda.time.DateTime;

import java.util.HashMap;
import java.util.Map;

import static com.cs.rfq.decorator.extractors.RfqMetadataFieldNames.*;

public class AverageTradedPriceExtractor implements RfqMetadataExtractor {
    private String since;
    public AverageTradedPriceExtractor() {
        this.since = DateTime.now().getYear() + "-01-01";
    }

    @Override
    public Map<RfqMetadataFieldNames, Object> extractMetaData(Rfq rfq, SparkSession session, Dataset<Row> trades) {

        long todayMs = DateTime.now().withMillisOfDay(0).getMillis();
        long pastWeekMs = DateTime.now().withMillis(todayMs).minusWeeks(1).getMillis();
        long pastYearMs = DateTime.now().withMillis(todayMs).minusYears(1).getMillis();

        Dataset<Row> filtered = trades
                .filter(trades.col("SecurityId").equalTo(rfq.getIsin()))
                .filter(trades.col("EntityId").equalTo(rfq.getEntityId()));

        filtered.createOrReplaceTempView("fil");

        String query = String.format("SELECT sum(LastPx)/count(LastPx) from fil");

        Dataset<Row> sqlQueryResults = session.sql(query);

        Object avgPx = sqlQueryResults.first().get(0);
        if (avgPx == null) {
            avgPx = 0L;
        }

        Map<RfqMetadataFieldNames, Object> results = new HashMap<>();
        results.put(RfqMetadataFieldNames.averageTradedPrice, avgPx);
        return results;
    }
    protected void setSince(String since) {
        this.since = since;
    }

}